Refactor to use interface and simplify PayPalApi implementation
